make precomputation_laplace as a separate compilation unit
