+++ Makefile
@@ -13,8 +13,8 @@ LDFLAGS = -lfftw3 -lfftw3f -lmkl_intel_lp64 -lmkl_intel_thread -lmkl_core -liomp
 #CXXFLAGS = -g -O3 -mavx -fabi-version=6 -std=c++11 -fopenmp -I./include
 #LDFLAGS = -lfftw3 -lfftw3f -lpthread -lblas -llapack -lm
 
-OBJF = main.fo src/geometry.fo src/laplace.fo
-OBJD = main.do src/geometry.do src/laplace.do
+OBJF = main.fo src/geometry.fo src/laplace.fo src/precompute_laplace.fo
+OBJD = main.do src/geometry.do src/laplace.do src/precompute_laplace.do
 OBJHF = main.hfo src/geometry.hfo src/helmholtz.hfo
 OBJHD = main.hdo src/geometry.hdo src/helmholtz.hdo
