+++ include/laplace.h
@@ -5,30 +5,17 @@
 #include "exafmm_t.h"
 #include "geometry.h"
 #include "intrinsics.h"
+#include "precompute_laplace.h"
 
 extern "C" {
-  void sgemm_(char* transa, char* transb, int* m, int* n, int* k, float* alpha, float* a,
-              int* lda, float* b, int* ldb, float* beta, float* c, int* ldc);
-  void dgemm_(char* transa, char* transb, int* m, int* n, int* k, double* alpha, double* a,
-              int* lda, double* b, int* ldb, double* beta, double* c, int* ldc);
   void sgemv_(char* trans, int* m, int* n, float* alpha, float* a, int* lda, float* x,
               int* incx, float* beta, float* y, int* incy);
   void dgemv_(char* trans, int* m, int* n, double* alpha, double* a, int* lda, double* x,
               int* incx, double* beta, double* y, int* incy);
-  void sgesvd_(char *jobu, char *jobvt, int *m, int *n, float *a, int *lda,
-               float *s, float *u, int *ldu, float *vt, int *ldvt, float *work, int *lwork, int *info);
-  void dgesvd_(char *jobu, char *jobvt, int *m, int *n, double *a, int *lda,
-               double *s, double *u, int *ldu, double *vt, int *ldvt, double *work, int *lwork, int *info);
 }
 
 namespace exafmm_t {
-  void gemm(int m, int n, int k, real_t* A, real_t* B, real_t* C);
-
   void gemv(int m, int n, real_t* A, real_t* x, real_t* y);
-
-  void svd(int m, int n, real_t* A, real_t* S, real_t* U, real_t* VT);
-
-  RealVec transpose(RealVec& vec, int m, int n);
 
   void potential_P2P(RealVec& src_coord, RealVec& src_value, RealVec& trg_coord, RealVec& trg_value);
