+++ src/laplace.cpp
@@ -6,17 +6,6 @@ namespace exafmm_t {
   int MAXLEVEL;
   M2LData M2Ldata;
 
-  //! blas gemm with row major data
-  void gemm(int m, int n, int k, real_t* A, real_t* B, real_t* C) {
-    char transA = 'N', transB = 'N';
-    real_t alpha = 1.0, beta = 0.0;
-#if FLOAT
-    sgemm_(&transA, &transB, &n, &m, &k, &alpha, B, &n, A, &k, &beta, C, &n);
-#else
-    dgemm_(&transA, &transB, &n, &m, &k, &alpha, B, &n, A, &k, &beta, C, &n);
-#endif
-  }
-
   //! blas gemv with row major data
   void gemv(int m, int n, real_t* A, real_t* x, real_t* y) {
     char trans = 'T';
@@ -27,36 +16,6 @@ namespace exafmm_t {
 #else
     dgemv_(&trans, &n, &m, &alpha, A, &n, x, &incx, &beta, y, &incy);
 #endif
-  }
-
-  //! lapack svd with row major data: A = U*S*VT, A is m by n
-  void svd(int m, int n, real_t* A, real_t* S, real_t* U, real_t* VT) {
-    char JOBU = 'S', JOBVT = 'S';
-    int INFO;
-    int LWORK = std::max(3*std::min(m,n)+std::max(m,n), 5*std::min(m,n));
-    LWORK = std::max(LWORK, 1);
-    int k = std::min(m, n);
-    RealVec tS(k, 0.);
-    RealVec WORK(LWORK);
-#if FLOAT
-    sgesvd_(&JOBU, &JOBVT, &n, &m, A, &n, &tS[0], VT, &n, U, &k, &WORK[0], &LWORK, &INFO);
-#else
-    dgesvd_(&JOBU, &JOBVT, &n, &m, A, &n, &tS[0], VT, &n, U, &k, &WORK[0], &LWORK, &INFO);
-#endif
-    // copy singular values from 1d layout (tS) to 2d layout (S)
-    for(int i=0; i<k; i++) {
-      S[i*n+i] = tS[i];
-    }
-  }
-
-  RealVec transpose(RealVec& vec, int m, int n) {
-    RealVec temp(vec.size());
-    for(int i=0; i<m; i++) {
-      for(int j=0; j<n; j++) {
-        temp[j*m+i] = vec[i*n+j];
-      }
-    }
-    return temp;
   }
 
   void potential_P2P(RealVec& src_coord, RealVec& src_value, RealVec& trg_coord, RealVec& trg_value) {
